+++ src/pinboard/tests.rs
@@ -57,6 +57,49 @@ fn test_set_cache_dir() {
 }
 
 #[test]
+fn find_tag_test() {
+    let _ = env_logger::try_init();
+    debug!("find_url_test: starting.");
+    let (_m1, _m2) = create_mockito_servers();
+    let mut _home = dirs::home_dir().unwrap();
+    _home.push(".cache");
+    _home.push("mockito-rusty-pin");
+    let cache_path = Some(_home);
+
+    let mut pinboard =
+        Pinboard::new(include_str!("api_token.txt"), cache_path).expect("Can't setup Pinboard");
+    pinboard.enable_fuzzy_search(false);
+
+    let r = pinboard.find_tag("timemachine");
+    assert!(r.is_ok());
+    let op = r.unwrap();
+    assert!(op.is_some());
+    let pins = op.unwrap();
+    assert_eq!(3, pins.len());
+
+    // Case insensitive search
+    let r = pinboard.find_tag("TimeMachine");
+    assert!(r.is_ok());
+    let op = r.unwrap();
+    assert!(op.is_some());
+    let pins = op.unwrap();
+    assert_eq!(3, pins.len());
+
+    let r = pinboard.find_tag("hacks");
+    assert!(r.is_ok());
+    let op = r.unwrap();
+    assert!(op.is_some());
+    let pins = op.unwrap();
+    assert_eq!(1, pins.len());
+
+    // Should find exact tags only
+    let r = pinboard.find_tag("hack");
+    assert!(r.is_ok());
+    let op = r.unwrap();
+    assert!(op.is_none());
+}
+
+#[test]
 fn find_url_test() {
     let _ = env_logger::try_init();
     debug!("find_url_test: starting.");
@@ -79,6 +122,7 @@ fn find_url_test() {
 
     assert_eq!(1, pins.len());
 
+    // find_url is case insensitive.
     let r =
         pinboard.find_url("http://blog.khubla.com/freebsd/time-machine-backups-using-FreeBSD-zfs");
     assert!(r.is_ok());